CtrlCore: Gtk: Fixed issue with close button
git-svn-id: svn://ultimatepp.org/upp/trunk@11025 f0d560ea-af0d-0410-9eb7-867de7ffcac7

+++ uppsrc/CtrlCore/GtkTop.cpp
@@ -196,7 +196,7 @@ void TopWindow::Minimize(bool effect)
 
 TopWindow& TopWindow::FullScreen(bool b)
 {
-	SetMode(b ? FULLSCREEN : OVERLAPPED);
+	SetMode(b ? (int)FULLSCREEN : (int)OVERLAPPED);
 	return *this;
 }
